order_consumer/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.4
Flask==1.1.1
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pathlib2==2.3.4
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.1
pytest-cov==2.7.1
six==1.12.0
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
psycopg2>=2.7,<3.0
kafka-python
redis
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Script
rq==1.2.2
